refactor: share chain-env scanner between client and server
The env-scan logic for RPC_URL_HTTP_<chainId> and NUXT_PUBLIC_SUBGRAPH_URI_<chainId> lived in two places: server/utils/enabled-chains.ts (used by server plugins and API handlers) and composables/useChainConfig.ts#scanEnv (used on the server branch of the universal composable). If the env convention ever changes, both have to be updated in sync.

Move the scanner to utils/chain-env.ts (universal, not server-only) so every caller imports from one source of truth. The composable drops its inline regex, the server plugins/handlers update their import paths. Also replace a leftover console.warn in entities/vault/factory.ts with logWarn to match the rest of entities/vault/.
