+++ composables/useChainConfig.ts
@@ -8,6 +8,7 @@
  * This avoids runtimeConfig (which is frozen in production) and works
  * with runtime-injected env vars (e.g. Doppler on Railway).
  */
+import { getEnabledChainIds, getSubgraphUris } from '~/utils/chain-env'
 
 interface ChainConfig {
   enabledChainIds: number[]
@@ -18,25 +19,8 @@ interface ChainConfig {
 let cached: ChainConfig | null = null
 
 function scanEnv(): ChainConfig {
-  const enabledChainIds: number[] = []
-
-  for (const [key, value] of Object.entries(process.env)) {
-    const rpcMatch = key.match(/^RPC_URL_HTTP_(\d+)$/)
-    if (rpcMatch && value) {
-      enabledChainIds.push(Number(rpcMatch[1]))
-    }
-  }
-
-  const subgraphUris: Record<string, string> = {}
-  for (const [key, value] of Object.entries(process.env)) {
-    const match = key.match(/^NUXT_PUBLIC_SUBGRAPH_URI_(\d+)$/)
-    if (match && value) {
-      subgraphUris[match[1]] = value
-    }
-  }
-
-  enabledChainIds.sort((a, b) => a - b)
-
+  const enabledChainIds = getEnabledChainIds()
+  const subgraphUris = getSubgraphUris()
   const enabledSet = new Set(enabledChainIds)
   const deprecatedChainIds = parseDeprecatedChains(process.env.DEPRECATED_CHAINS, enabledSet)
